+++ sphinx/source/changelog.rst
@@ -6,8 +6,9 @@ Changelog
 * Dropped support for Python 2.7 and 3.3
 * Added the ``--expand`` option to :ref:`sna2img.py` (for expanding a
   ``#FONT``, ``#SCR``, ``#UDG`` or ``#UDGARRAY`` macro)
-* Added the ``--find-tile`` option to :ref:`snapinfo.py` (for searching for the
-  graphic data of a tile currently on screeen)
+* Added the ``--find-tile`` option (for searching for the graphic data of a
+  tile currently on screen) and the ``--word`` option (for showing the words at
+  a range of addresses) to :ref:`snapinfo.py`
 * Added support to the ``--find`` option of :ref:`snapinfo.py` for specifying a
   range of distances between byte values (e.g. ``--find 1,2,3-1-10``)
 * Added support for parsing a :ref:`box page <boxpages>` entry section as a